[kernel] cputime: Fix jiffies based cputime assumption on steal accounting
Message-id: <1396310684-6099-4-git-send-email-fweisbec@redhat.com>
Patchwork-id: 78747
O-Subject: [RHEL7 PATCH 3/3] cputime: Fix jiffies based cputime assumption on steal accounting
Bugzilla: 1047732
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Bz: 1047732

Commit dee08a72deefac251267ed2717717596aa8b6818
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Wed, 5 Mar 2014 17:02:22 +0100

The steal guest time accounting code assumes that cputime_t is based on
jiffies. So when CONFIG_NO_HZ_FULL=y, which implies that cputime_t
is based on nsecs, steal_account_process_tick() passes the delta in
jiffies to account_steal_time() which then accounts it as if it's a
value in nsecs.

As a result, accounting 1 second of steal time (with HZ=100 that would
be 100 jiffies) is spuriously accounted as 100 nsecs.

As such /proc/stat may report 0 values of steal time even when two
guests have run concurrently for a few seconds on the same host and
same CPU.

In order to fix this, lets convert the nsecs based steal delta to
cputime instead of jiffies by using the right conversion API.

Given that the steal time is stored in cputime_t and this type can have
a smaller granularity than nsecs, we only account the rounded converted
value and leave the remaining nsecs for the next deltas.

Reported-by: Huiqingding <huding@redhat.com>
Reported-by: Marcelo Tosatti <mtosatti@redhat.com>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Marcelo Tosatti <mtosatti@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Acked-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Frederic Weisbecker <fweisbec@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>