[redhat] kabi: Remove GENKSYMS usage through the code
Message-id: <1396633698-9231-2-git-send-email-dzickus@redhat.com>
Patchwork-id: 78796
O-Subject: [RHEL7 PATCH 1/2 V2] kabi: Remove GENKSYMS usage through the code
Bugzilla: 1083718
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1083718
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7305999
Tested:  Compile testing is the most important.  Boot tested for sanity
Upstream: RHEL only

A conversation started about the usage of GENKSYMS on a 7.0 kernel.

Normally this is not needed because kabi is not frozen until _after_ a .0
release is GA.

However folks were using this wrapper to clearly indicate which chunks of code
were being reserved for Red Hat purposes.

After thinking about this and discussing it with folks, we realized this is not
a good idea.  Two reasons have stuck out as to why.

First, the idea was that hiding the symbols in the GENKSYSMS wrapper, we hide
the true checksum of the symbols.  We should try to encompass the true checksum
for GA.  On the flip side, removing the macro and doing things properly for 7.1
would entail:

struct foo {
	int bar;
 #ifndef __GENKSYMS__
	int original;
 #else
	int new;
 #endif
}

which leads to ugly and hard to read code.  I agreed to create some macros in
7.1 that would hide the ugliness, yet make it obvious what is going on.

Second, and this is more a social problem, most folks have come to assume that
anything inside __GENKSYMS__ is free to be modiifed/added.  But in this
particular example, this is _not_ the case.  So to prevent the confusion, we
should avoid this usage for reserving memory.

As a result, I have removed all uses I believe were not correct.  There should
be no downside as this only affects how genksyms pre-processes the symbols and
generates checksums.  The only noticable change will be new kabi checksums for
a bunch of functions.

I did leave in a few usages, namely the PCI ones.  I believe this is a correct
usage as it properly sets the stage for future additions to the reserved
struct.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>