[redhat] Kconfig: Make MEI modular
Message-id: <1397566280-6623-2-git-send-email-prarit@redhat.com>
Patchwork-id: 78906
O-Subject: [RHEL7.1 PATCH BZ 1082789 1/2] Make MEI modular
Bugzilla: 1082789
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Michal Schmidt <mschmidt@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1082789

MEI is only used on some desktops and some laptops.  There is no risk
making it a module and it is a module in Fedora.  I was planning on
pushing this for 7.1 anyway, and doing so now will minimize the impact to
our customers in 7.1.

This hopefully will resolve any further issues that our customers have
with the MEI as they will be able to blacklist it from loading.

Successfully tested by me.

Cc: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>