[fs] autofs: fix lockref lookup
Message-id: <20140429011306.5928.25984.stgit@perseus.fritz.box>
Patchwork-id: 79218
O-Subject: [RHEL-7.0 PATCH] autofs - fix lockref lookup
Bugzilla: 1083815
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Bugzilla bug: https://bugzilla.redhat.com/show_bug.cgi?id=1083815
Upstream commit: not yet committed, included in mm tree 04/23/2014

autofs needs to be able to see private data dentry flags for
its dentrys that are being created but not yet hashed and for
its dentys that have been rmdir()ed but not yet freed. It
needs to do this so it can block processes in these states
until a status has been returned to indicate the given
operation is complete.

It does this by keeping two lists, active and expring, of
dentrys in this state and uses ->d_release() to keep them
stable while it checks the reference count to determine
if they should be used.

But with the recent lockref changes dentrys being freed
sometimes don't transition to a reference count of 0 before
being freed so autofs can occassionally use a dentry that
is invalid which can lead to a panic.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Jarod Wilson <jarod@redhat.com>