[virt] hyperv/vmbus: Negotiate version 3.0 when running on ws2012r2 hosts
Message-id: <1399395698-10215-2-git-send-email-vgoyal@redhat.com>
Patchwork-id: 79414
O-Subject: [RHEL7.0 PATCH 1/2] Drivers: hv: vmbus: Negotiate version 3.0 when running on ws2012r2 hosts
Bugzilla: 1037957
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jason Wang <jasowang@redhat.com>

BZ:1037957

This is backport of upstream commit 03367ef5ea811475187a0732aada068919e14d61

Only ws2012r2 hosts support the ability to reconnect to the host on VMBUS. This functionality
is needed by kexec in Linux. To use this functionality we need to negotiate version 3.0 of the
VMBUS protocol.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Cc: <stable@vger.kernel.org>        [3.9+]
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>