[fs] gfs2: Increase the max number of ACLs
Message-id: <451511823.9721934.1400693350822.JavaMail.zimbra@redhat.com>
Patchwork-id: 80073
O-Subject: [RHEL7.1 PATCH][bz1078231] GFS2: Increase the max number of ACLs
Bugzilla: 1078231
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This is a RHEL7.1 back-port of a patch that's already gone into other
releases. The upstream patch is here:

http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=b00263d1cafdd667de56cde47f35d5ee8dd37e14

It was tested on system gfs-i24c-01.mpc.lab.eng.bos.redhat.com with
a reliable reproducer.

This patch increases the maximum number of ACLs from 25 to 300 for
a 4K block size. The value is adjusted accordingly if the block size
is smaller. Note that this is an arbitrary limit with a performance
tradeoff, and that the physical limit is slightly over 500.

rhbz#1078231

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>