[mm] avoid kABI breakage
Message-id: <1401356312-17063-6-git-send-email-jmarchan@redhat.com>
Patchwork-id: 81104
O-Subject: [RHEL7 PATCH 5/5] avoid kABI breakage
Bugzilla: 1097419
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=1097419

The changes in swap_info_struct breaks the kABI. This is because one
argument of the (mostly unused) swap_activate() prototype of
address_space_operations is of that type. The struct address_space is
then refered to by several other structures and the modification is
eventually transmitted to a few whitelisted symbols unrelated to the
modified swap_info_struct structure.

The structure is never included directly in other structure, so hiding
the new fields at the end should be fine.

Signed-off-by: Jarod Wilson <jarod@redhat.com>