[block] mtip32xx: Unmap the DMA segments before completing the IO request
Message-id: <1401397882-5166-2-git-send-email-jmoyer@redhat.com>
Patchwork-id: 81191
O-Subject: [RHEL7 PATCH 1/2] [block] mtip32xx: Unmap the DMA segments before completing the IO request
Bugzilla: 1102281
RH-Acked-by: David Milburn <dmilburn@redhat.com>

This is a backport of the following commit:

  commit 368c89d7ac70f937c93cd6f3b65bcfdfb3ba794f
  Author: Felipe Franciosi <felipe@paradoxo.org>
  Date:   Thu Mar 13 14:34:21 2014 +0000

    If the buffers are unmapped after completing a request, then stale data
    might be in the request.

    Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>
    Cc: stable@kernel.org
    Signed-off-by: Jens Axboe <axboe@fb.com>

This patch is required for the next one to apply cleanly, but is also
important enough to include in z-stream.  Micron has agreed to test
the fix, as we don't have the hardware in our lab.

brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7512189
bz: 1102281

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>