[fs] xfs: xfs_readsb needs to check for magic numbers
Message-id: <5387A74A.3020107@redhat.com>
Patchwork-id: 81193
O-Subject: [RHEL7.1 PATCH] xfs: xfs_readsb needs to check for magic numbers
Bugzilla: 1102940
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>

This is for:

Bug 1102940 - fs: xfs_readsb needs to check for magic numbers during probe-based mounts (edit)

The following patch is on the xfs list & reviewed, but not yet upstream:

xfs: xfs_readsb needs to check for magic numbers

From: Dave Chinner <dchinner@redhat.com>

Commit daba542 ("xfs: skip verification on initial "guess"
superblock read") dropped the use of a verifier for the initial
superblock read so we can probe the sector size of the filesystem
stored in the superblock. It, however, now fails to validate that
what was read initially is actually an XFS superblock and hence will
fail the sector size check and return ENOSYS.

This causes probe-based mounts to fail because it expects XFS to
return EINVAL when it doesn't recognise the superblock format.

cc: <stable@vger.kernel.org>
Reported-by: Plamen Petrov <plamen.sisi@gmail.com>
Tested-by: Plamen Petrov <plamen.sisi@gmail.com>
Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>