[redhat] kernel.spec: make weak-updates actually work
Message-id: <1401453715-18059-1-git-send-email-jarod@redhat.com>
Patchwork-id: 81241
O-Subject: [RHEL7 PATCH] [redhat] kernel.spec: make weak-updates actually work
Bugzilla: 1101036
RH-Acked-by: Phillip Lougher <plougher@redhat.com>

Bugzilla: 1101036

Somehow, we neglected to notice that the weak-updates infrastructure never
made its way into the RHEL7 kernel. This restores the functionality, with
special thanks to Akemi Yagi of elrepo.org for discovering the problem and
an initial patch.

This is going to have to be added to z-stream as well, or kmod updates
built against the GA kernel won't work with any other 7.0.z kernel.

CC: Phillip Lougher <plougher@redhat.com>
CC: Don Howards <dhoward@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>