[fs] gfs2: Prevent recovery before the local journal is set
Message-id: <1521928274.16571189.1401884740283.JavaMail.zimbra@redhat.com>
Patchwork-id: 82432
O-Subject: [RHEL7.1 PATCH][bz #1079433] GFS2: Prevent recovery before the local journal is set
Bugzilla: 1079433
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bz #1079433 - GFS2: NULL pointer in gfs2_recover_set.
The symptom: a node is fenced and when it returns, its journal is replayed,
but the timing is such that it tries to recover the journal before the
journal is opened, which results in a kernel panic. This patch fixes the
problem by using a completion to wait for journal initialization before
attempting journal recovery.

The upstream patch is here:
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=0e48e055a7dfc0cf17bbabe4d9b523ee0b1a9ed6

A test kernel was built here:
https://brewweb.devel.redhat.com/taskinfo?taskID=7488867

It was tested on a four-node cluster using the "revolver" QE test on nodes:
gfs-i8c-0{1..4}.mpc.lab.eng.bos.redhat.com.

Patch description:

This patch uses a completion to prevent dlm's recovery process from
referencing and trying to recover a journal before a journal has been
opened.

rhbz#1079433

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>