[mm] hugetlb: ensure hugepage access is denied if hugepages are not supported
Message-id: <20140616183422.e465f63c825efbf1dc014eef@redhat.com>
Patchwork-id: 84000
O-Subject: [RHEL7 PATCH] hugetlb: ensure hugepage access is denied if hugepages are not supported
Bugzilla: 1081671
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>

Bugzilla: 1081671

Notes:
- upstream commit 457c1b27ed56ec472d202731b12417bff023594a "hugetlb:
  ensure hugepage access is denied if hugepages are not supported"
  fixes a crash on RHEL running as a guest under KVM on Power, when
  the guest is not backed by host hugepages.

Currently, I am seeing the following when I `mount -t hugetlbfs /none
/dev/hugetlbfs`, and then simply do a `ls /dev/hugetlbfs`.  I think it's
related to the fact that hugetlbfs is properly not correctly setting
itself up in this state?:

  Unable to handle kernel paging request for data at address 0x00000031
  Faulting instruction address: 0xc000000000245710
  Oops: Kernel access of bad area, sig: 11 [#1]
  SMP NR_CPUS=2048 NUMA pSeries
  ....

In KVM guests on Power, in a guest not backed by hugepages, we see the
following:

  AnonHugePages:         0 kB
  HugePages_Total:       0
  HugePages_Free:        0
  HugePages_Rsvd:        0
  HugePages_Surp:        0
  Hugepagesize:         64 kB

HPAGE_SHIFT == 0 in this configuration, which indicates that hugepages
are not supported at boot-time, but this is only checked in
hugetlb_init().  Extract the check to a helper function, and use it in a
few relevant places.

This does make hugetlbfs not supported (not registered at all) in this
environment.  I believe this is fine, as there are no valid hugepages
and that won't change at runtime.

[akpm@linux-foundation.org: use pr_info(), per Mel]
[akpm@linux-foundation.org: fix build when HPAGE_SHIFT is undefined]
Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>
Reviewed-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Acked-by: Mel Gorman <mgorman@suse.de>
Cc: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 457c1b27ed56ec472d202731b12417bff023594a)

Signed-off-by: Jarod Wilson <jarod@redhat.com>