[drm] i915: make i915 less chatty on resume
Message-id: <1403093381-31787-1-git-send-email-rclark@redhat.com>
Patchwork-id: 84256
O-Subject: [RHEL7 drm] drm/i915: make i915 less chatty on resume
Bugzilla: 1086777
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Dave Airlie <airlied@redhat.com>
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>

RHBZ: #1086777
Upstream: n/a

We hit these WARN_ON()'s on resume from suspend, although they are
benign.

The issue is known upstream, and a proper fix is being worked on.  But
that will be more of a refactoring, and not something we can pull in
before rhel71.  So to cut down on spurrious bug reports, remove the
WARN_ON()'s.

Signed-off-by: Rob Clark <rclark@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>