[fs] nfsd: fh_update should error out in unexpected cases
Message-id: <1403560314-32525-9-git-send-email-bfields@redhat.com>
Patchwork-id: 84446
O-Subject: [Patch RHEL7 008/172] nfsd: fh_update should error out in unexpected cases
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

The reporter saw a NULL dereference when a filesystem's ->mknod returned
success but left the dentry negative, and then nfsd tried to dereference
d_inode (in this case because the CREATE was followed by a GETATTR in
the same nfsv4 compound).

fh_update already checks for this and another broken case, but for some
reason it returns success and leaves nfsd trying to soldier on.  If it
failed we'd avoid the crash.  There's only so much we can do with a
buggy filesystem, but it's easy enough to bail out here, so let's do
that.

Reported-by: Antti Tönkyrä <daedalus@pingtimeout.net>
Tested-by: Antti Tönkyrä <daedalus@pingtimeout.net>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 49e7372063a220651b4f12fa46113559533cafee
Bugzilla: 1111719
---
 fs/nfsd/nfsfh.c | 8 +++-----
 1 file changed, 3 insertions(+), 5 deletions(-)

Signed-off-by: Jarod Wilson <jarod@redhat.com>