[fs] nfsd: -EINVAL on invalid anonuid/gid instead of silent failure
Message-id: <1403560314-32525-11-git-send-email-bfields@redhat.com>
Patchwork-id: 84464
O-Subject: [Patch RHEL7 010/172] nfsd: -EINVAL on invalid anonuid/gid instead of silent failure
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

If we're going to refuse to accept these it would be polite of us to at
least say so....

This introduces a slight complication since we need to grandfather in
exportfs's ill-advised use of -1 uid and gid on its test_export.

If it turns out there are other users passing down -1 we may need to
do something else.

Best might be to drop the checks entirely, but I'm not sure if other
parts of the kernel might assume that a task can't run as uid or gid -1.

Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 6f6cc3205c5f10129b8a10cdf8abf85d9db48a60
Bugzilla: 1111719
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>