[fs] nfsd: improve write performance with better sendspace reservations
Message-id: <1403560314-32525-16-git-send-email-bfields@redhat.com>
Patchwork-id: 84452
O-Subject: [Patch RHEL7 015/172] nfsd4: improve write performance with better sendspace reservations
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Currently the rpc code conservatively refuses to accept rpc's from a
client if the sum of its worst-case estimates of the replies it owes
that client exceed the send buffer space.

Unfortunately our estimate of the worst-case reply for an NFSv4 compound
is always the maximum read size.  This can unnecessarily limit the
number of operations we handle concurrently, for example in the case
most operations are writes (which have small replies).

We can do a little better if we check which ops the compound contains.

This is still a rough estimate, we'll need to improve on it some day.

Reported-by: Shyam Kaushik <shyamnfs1@gmail.com>
Tested-by: Shyam Kaushik <shyamnfs1@gmail.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 6ff40decff0ef35a5d755ec60182d7f803356dfb
Bugzilla: 1111719
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>