[fs] nfsd: decrease nfsd4_encode_fattr stack usage
Message-id: <1403560314-32525-42-git-send-email-bfields@redhat.com>
Patchwork-id: 84473
O-Subject: [Patch RHEL7 041/172] nfsd4: decrease nfsd4_encode_fattr stack usage
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

A struct svc_fh is 320 bytes on x86_64, it'd be better not to have these
on the stack.

kmalloc'ing them probably isn't ideal either, but this is the simplest
thing to do.  If it turns out to be a problem in the readdir case then
we could add a svc_fh to nfsd4_readdir and pass that in.

Acked-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: d50e61361c68a05a9cd7d54617522f99f278ac8a
Bugzilla: 1111719
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>