[net] svcrpc: explicitly reject compounds that are not padded out to 4-byte multiple
Message-id: <1403560314-32525-51-git-send-email-bfields@redhat.com>
Patchwork-id: 84461
O-Subject: [Patch RHEL7 050/172] svcrpc: explicitly reject compounds that are not padded out to 4-byte multiple
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

We have a WARN_ON in the nfsd4_decode_write() that tells us when the
client has sent a request that is not padded out properly according to
RFC4506. A WARN_ON really isn't appropriate in this case though since
this indicates a client bug, not a server one.

Move this check out to the top-level compound decoder and have it just
explicitly return an error. Also add a dprintk() that shows the client
address and xid to help track down clients and frames that trigger it.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: e874f9f8e04cb67351893894dfb9fbcd25e62fa2
Bugzilla: 1111719
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>