[fs] nfsd: Only set PF_LESS_THROTTLE when really needed
Message-id: <1403560314-32525-96-git-send-email-bfields@redhat.com>
Patchwork-id: 84611
O-Subject: [Patch RHEL7 095/172] nfsd: Only set PF_LESS_THROTTLE when really needed.
Bugzilla: 1111719
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: NeilBrown <neilb@suse.de>

PF_LESS_THROTTLE has a very specific use case: to avoid deadlocks
and live-locks while writing to the page cache in a loop-back
NFS mount situation.

It therefore makes sense to *only* set PF_LESS_THROTTLE in this
situation.
We now know when a request came from the local-host so it could be a
loop-back mount.  We already know when we are handling write requests,
and when we are doing anything else.

So combine those two to allow nfsd to still be throttled (like any
other process) in every situation except when it is known to be
problematic.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 8658452e4a588da603f6cb5ee2615deafcd82b71
Bugzilla: 1111719
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>