[fs] nfs: modify pg_test interface to return size_t
Message-id: <1403798077-28799-29-git-send-email-steved@redhat.com>
Patchwork-id: 84941
O-Subject: [RHEL7.1 PATCH 28/78] nfs: modify pg_test interface to return size_t
Bugzilla: 1113248
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Weston Andros Adamson <dros@primarydata.com>

This is a step toward allowing pg_test to inform the the
coalescing code to reduce the size of requests so they may fit in
whatever scheme the pg_test callback wants to define.

For now, just return the size of the request if there is space, or 0
if there is not.  This shouldn't change any behavior as it acts
the same as when the pg_test functions returned bool.

Signed-off-by: Weston Andros Adamson <dros@primarydata.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id: b4fdac1a5150174df0847a45dc6612ce5ce3daeb
Bugzilla: 1113248
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>