[fs] nfs: Comment is now wrong and redundant to code
Message-id: <1403798077-28799-76-git-send-email-steved@redhat.com>
Patchwork-id: 84920
O-Subject: [RHEL7.1 PATCH 75/78] NFSv4.1: Comment is now wrong and redundant to code
Bugzilla: 1113248
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Tom Haynes <Thomas.Haynes@primarydata.com>

The save of the write offset was removed some time ago, so that
part of the comment is bogus.

The remainder is pretty self-evident.

So off with it!

Signed-off-by: Tom Haynes <Thomas.Haynes@primarydata.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id:
Bugzilla: 1113248
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>