[fs] gfs2: Fix race in glock lru glock disposal
Message-id: <4a3782265178bf949466917f196e5a7bc3a47ce2.1404245053.git.rpeterso@redhat.com>
Patchwork-id: 85170
O-Subject: [RHEL7.1 PATCH 2/2] GFS2: Fix race in glock lru glock disposal
Bugzilla: 1095835
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

We must not leave items on the LRU list with GLF_LOCK set, since
they can be removed if the glock is brought back into use, which
may then potentially result in a hang, waiting for GLF_LOCK to
clear.

It doesn't happen very often, since it requires a glock that has
not been used for a long time to be brought back into use at the
same moment that the shrinker is part way through disposing of
glocks.

The fix is to set GLF_LOCK at a later time, when we already know
that the other locks can be obtained. Also, we now only release
the lru_lock in case a resched is needed, rather than on every
iteration.

rhbz#1095835

Signed-off-by: Jarod Wilson <jarod@redhat.com>