[redhat] spec: hide changelog lines for unsupported arches
Message-id: <20140715205551.GA8007@redhat.com>
Patchwork-id: 85841
O-Subject: [RHEL7 PATCH RFC] redhat/spec: hide changelog lines for unsupported arches
Bugzilla:
RH-Acked-by: Don Zickus <dzickus@redhat.com>

This is a quick hack to genspec.sh that more or less duplicates the logic
used to suppress the [redhat] commits from the public kernel changelog
stashed in the end-result kernel rpms, to hide any commits that start with
[xyz], where xyz is an arch we don't support. Pretty straight-forward
stuff, and should niftily cope with one of the major objections to us
merging patches that touch unsupported arches. If its not in the rpm
changelog, it didnt' happen, right?

List created by an ls of arch/ in the kernel, omitting Kconfig, x86, s390
and powerpc, for obvious reasons.

---

Signed-off-by: Jarod Wilson <jarod@redhat.com>