[virt] kvm/x86: add comments where MMIO does not return to the emulator
Message-id: <1406040016-3289-75-git-send-email-pbonzini@redhat.com>
Patchwork-id: 86153
O-Subject: [RHEL7 PATCH v2 074/221] KVM: x86: add comments where MMIO does not return to the emulator
Bugzilla: 1116936
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Support for single-step in the emulator (new in 3.12) does not work for
MMIO or PIO writes, because they are completed without returning to
the emulator.  This is not worse than what we had in 3.11; still, add
comments so that the issue is not forgotten.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Gleb Natapov <gleb@redhat.com>
(cherry picked from commit 0912c9771e9902f752e890e93af495cc06a786ac)
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>