[virt] kvm/x86: Allow the guest to run with dirty debug registers
Message-id: <1406040016-3289-162-git-send-email-pbonzini@redhat.com>
Patchwork-id: 86240
O-Subject: [RHEL7 PATCH v2 161/221] KVM: x86: Allow the guest to run with dirty debug registers
Bugzilla: 1068627 1116936
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When not running in guest-debug mode, the guest controls the debug
registers and having to take an exit for each DR access is a waste
of time.  If the guest gets into a state where each context switch
causes DR to be saved and restored, this can take away as much as 40%
of the execution time from the guest.

After this patch, VMX- and SVM-specific code can set a flag in
switch_db_regs, telling vcpu_enter_guest that on the next exit the debug
registers might be dirty and need to be reloaded (syncing will be taken
care of by a new callback in kvm_x86_ops).  This flag can be set on the
first access to a debug registers, so that multiple accesses to the
debug registers only cause one vmexit.

Note that since the guest will be able to read debug registers and
enable breakpoints in DR7, we need to ensure that they are synchronized
on entry to the guest---including DR6 that was not synced before.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit c77fb5fe6f031bee9403397ae7b94ea22ea19aa7)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>